refactor(6895): tighten group channel crypto regression support
